apiserver: don't Kill presence if mongo is dead
When mongo is unavailable/unreachable, don't attempt
to "kill" the agent presence. Presence is informational,
and so updating it should not block connection termination
when Mongo is unreachable.

Fixes https://bugs.launchpad.net/juju-core/+bug/1497229

+++ apiserver/admin.go
@@ -310,6 +310,7 @@ func checkForValidMachineAgent(entity state.Entity, req params.LoginRequest) err
 // machinePinger wraps a presence.Pinger.
 type machinePinger struct {
 	*presence.Pinger
+	mongoUnavailable *bool
 }
 
 // Stop implements Pinger.Stop() as Pinger.Kill(), needed at
@@ -318,7 +319,15 @@ func (p *machinePinger) Stop() error {
 	if err := p.Pinger.Stop(); err != nil {
 		return err
 	}
-	return p.Pinger.Kill()
+	if !*p.mongoUnavailable {
+		// Kill marks the agent as not-present. If the
+		// Mongo server is known to be unavailable, then
+		// we do not perform this operation; the agent
+		// will naturally become "not present" when its
+		// presence expires.
+		return p.Pinger.Kill()
+	}
+	return nil
 }
 
 func startPingerIfAgent(root *apiHandler, entity state.Entity) error {
@@ -336,7 +345,7 @@ func startPingerIfAgent(root *apiHandler, entity state.Entity) error {
 		return err
 	}
 
-	root.getResources().Register(&machinePinger{pinger})
+	root.getResources().Register(&machinePinger{pinger, root.mongoUnavailable})
 	action := func() {
 		if err := root.getRpcConn().Close(); err != nil {
 			logger.Errorf("error closing the RPC connection: %v", err)
